chore(phase-14): lint / type / DX — expand ruff ruleset, Makefile, mypy strictness
Move from a minimal lint config (E/F/I/UP only) to a ruleset that
catches likely bugs and stale code patterns, and turn on a first wave
of mypy strictness without requiring a repo-wide annotation sweep.

Ruff additions
* Enabled B (flake8-bugbear), RUF (ruff-native), SIM (flake8-simplify)
  and C4 (comprehensions) on top of the existing E/F/I/UP.
* Narrow per-file ignores:
  - tests/** drop S101 / ANN / SIM105 / RUF043 (idiomatic in test
    scaffolding).
  - cli/** drop B008 (typer.Option/Argument defaults) and B904
    (CLI handlers convert to `typer.Exit(1)` and hide tracebacks
    from end users by design).
  - api/runs.py + api/optimization.py + engine/two_stage.py keep B008
    (FastAPI `Query(...)` / library factory defaults).
  - asrbench/static/** lints disabled (generated Vite bundle).
* Repo-wide ignores: RUF001/002/003 (unicode em-dashes are
  deliberate), SIM102 / SIM113 (nested-if collapses and the
  enumerate rewrite aren't always clearer).
* Applied ~60 auto/unsafe fixes: contextlib.suppress over
  try/except/pass, `raise ... from exc` on exceptions crossing a
  boundary, `zip(..., strict=False)` to keep the 3.10+ behaviour
  explicit, sorted `__all__`, deleted unused `# noqa` directives,
  ClassVar for shared frozenset attributes.
* 2 fire-and-forget `create_task` call sites keep their task
  reference dropped (activity logger, VRAM sampler) with an
  inline `# noqa: RUF006` plus a why-comment.

Mypy
* Added warn_unused_configs / warn_unused_ignores /
  warn_redundant_casts / warn_return_any / strict_equality /
  no_implicit_optional. Legacy `strict=false` retained so the
  repo-wide annotation sweep can land in a later phase.
* Per-module override: tests keep untyped helpers.

DX scaffolding
* `Makefile` (~100 lines) — `install`, `lint`, `format`, `typecheck`,
  `test`, `test-cov`, `coverage-html`, `ui-build`, `ui-dev`,
  `pre-commit`, `docker-build`, `clean`. Every target prefers
  `uv run` when available and falls back to plain `python` /
  `pip` / `npm` so it works on a fresh clone.
* `.python-version` pinned to 3.12 (the CI reference leg).
* `CONTRIBUTING.md` gets an env-var reference table (ASRBENCH_API_KEY,
  ASRBENCH_ALLOWED_PATHS, HF_TOKEN, CUDA_VISIBLE_DEVICES) so new
  contributors can see the full surface without reading code.

Regression: 619 passed / 4 skipped / 0 failed. Coverage 72.85 %
(held above the 70 % gate).

+++ asrbench/engine/search/local_1d.py
@@ -261,7 +261,7 @@ def golden_section_search(
     # Adaptive patience: when the budget is nearly used up, stop sooner.
     # Scale linearly between 1 and 3 based on remaining fraction.
     remaining_ratio = budget.remaining / budget.hard_cap if budget.hard_cap else 1.0
-    adaptive_patience = max(1, int(round(3 * remaining_ratio)))
+    adaptive_patience = max(1, round(3 * remaining_ratio))
 
     for i in range(max_iterations):
         iterations = i + 1
@@ -447,7 +447,7 @@ def pattern_search(
     # cap the outer loop more tightly so we don't burn the remainder on a
     # step-halving cascade that won't escape.
     remaining_ratio = budget.remaining / budget.hard_cap if budget.hard_cap else 1.0
-    effective_max = max(1, int(round(max_iterations * max(remaining_ratio, 0.25))))
+    effective_max = max(1, round(max_iterations * max(remaining_ratio, 0.25)))
 
     for i in range(effective_max):
         iterations = i + 1
